+++ x/distribution/client/cli/utils.go
@@ -1,26 +1,13 @@
-package utils
+package cli
 
 import (
 	"io/ioutil"
 
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/types/rest"
 )
 
 type (
-	// CommunityPoolSpendProposalReq defines a community pool spend proposal request body.
-	CommunityPoolSpendProposalReq struct {
-		BaseReq rest.BaseReq `json:"base_req"`
-
-		Title       string         `json:"title"`
-		Description string         `json:"description"`
-		Recipient   sdk.AccAddress `json:"recipient"`
-		Amount      sdk.Coins      `json:"amount"`
-		Proposer    sdk.AccAddress `json:"proposer"`
-		Deposit     sdk.Coins      `json:"deposit"`
-	}
-
 	// CommunityPoolSpendProposalJSON defines a CommunityPoolSpendProposal with a deposit
 	CommunityPoolSpendProposalJSON struct {
 		Title       string         `json:"title"`
